extract and add test for getFlags
