+++ commands/slash_handler.js
@@ -1,4 +1,5 @@
 const success = require('./success')
+const getFlags = require('./get_flags')
 
 const VERIFY_TOKEN = process.env.verificationToken
 
@@ -273,55 +274,6 @@ function getChannelsWithFlags(controller, bot, message, cb) {
   }
 }
 
-function getFlags(controller, bot, message, cb) {
-  var sendbackTeamChannels = function(err, channels) {
-    // This allows us to set the default count for a given flag to 0
-    var flagDict =  new Proxy({}, {
-                      get: function(object, property) {
-                        return object.hasOwnProperty(property) ? object[property] : 0;
-                      }
-                    });
-    if (!err && channels) {
-      channels.map(function(c) {
-        // This conditional may seem redundant for .find() cases
-        // but see AUDIT note below
-        if (c.team_id == message.team_id && c.label) {
-          flagDict[c.label] += 1;
-        }
-      })
-    }
-    
-    var finalMessage = flagFormatting(flagDict);
-
-    cb(err, finalMessage);
-  }
-  var storageChannels = controller.storage.channels
-
-  if (storageChannels.find) {
-    // not all storage backends have find()
-    // e.g. Mongodb has it, but redis does not
-    storageChannels.find({team_id: message.team_id}, sendbackTeamChannels)
-  } else {
-    // AUDIT NOTE: This channels.all gets all channels across
-    // all instances -- not just the team instance
-    // however you'll see we filter on message.team_id matching above
-    // so nothing leaks (efficiency may be another question).
-    storageChannels.all(sendbackTeamChannels)
-  }
-}
-
-function flagFormatting(flagCounts) {
-  var formattedList = '```' +"All Flags:\n";
-
-  for (var flag in flagCounts) {
-    formattedList += flag;
-    formattedList += "\n";
-  }
-
-  formattedList += '```';
-  return formattedList
-}
-
 function logOut(controller, bot, message){
   let user = message.user_id
 
@@ -394,7 +346,7 @@ publicMethods.mainHandler = function(controller, bot, message) {
       break
     case '/getflags':
       // list all the flags
-      getFlags(controller, bot, message, function(err, flags) {
+      getFlags.call(controller, bot, message, function(err, flags) {
         bot.replyPrivate(message, flags)
       })
       break
